requirements/runtime.txt
addict
func_timeout
numpy
openapi-pydantic
packaging
Pillow
prompt_toolkit
pydantic>=2.0
pyyaml
requests
rich
thefuzz
tqdm
typing-extensions
